refactor project to PyCharm's
