--- colors.py
@@ -1,7 +1,0 @@
-colors = {
-    "WHITE": [255, 255, 255],
-    "BLACK": [21, 29, 33],
-    "RED": [255, 0, 0],
-    "BLUE" : [0, 0, 255]
-
-}
