--- docs/ideas.txt
@@ -1,39 +1,0 @@
-
-# How to make the DRAWINING of player change?
-
-# How to check if the board array is full, and which parts are AVAILABLE?
-Checking if the spot is available avoids the player override the play of your opponent (which is unfair :P)
-
-todos os espaços do jogo podem estar tanto DISPONIVEIS, quanto INDISPONIVEIS pois pode já haver algum desenho nele.
-a tecnica pra checar se já existe um desenho nele nao deu certo entao temos que pensar em outra forma
-THE CODING TRAIN HELP
-
-criar um array com as posições disponiveis e remover a posição em que o mouse clicou
-no video ele cria uma funçao nextTurn, no meu codigo eu deveria fazer isso na função play()
-se nao tiver mais lugares disponiveis o array available deve estar COMPLETO (length = 0) e sem nenhum vencedor (criar uma variavel pra o vencedor)
-
-- fazer tudo isso na classe player ou na classe board?
-como na classe board tem a função play() entao é possivel fazer por lá
-
-
-# How to check if someone wins? And how to check the directions that the player won (diagonals, etc)
-To check that there was a tie, the board must be full AND without any wins.
-To someone win, the board can be full or not (because it could be the last play).
-So there are two diferent conditions:
-- if the board is full/there's no spots available AND no wins then: tie
-- if he have some win : the player wins
-
-checar as horizontais, verticais e diagonais
-    cada um precisa ser igual, MAS NAO PODEM SER VAZIOS. pois se forem vazios o jogo vai pensar que a parte branca ('') ganhou
-
-
-fazer uma versao que o computador joge sozinho kkkkk ia ser louco
-----
-
-pra fazer o esquema de checagem na outra diagonal, é preciso analisar a logica por trás:
-    - um index cresce e o outro diminui
-    - um começa do 0 e o outro começa com o length da variavel board. e assim ele vai diminuindo e o outro aumentando até que os valores deles coincidem (quando chega no meio do tabuleiro)
-    - fazer usando for loops separados, nao faz sentido ser junto
-mas isso ai ta uma pora ;-;
-
-até agora eu nao sei se eu faço esse txt em ingles ou portugues lolololol
