--- docs/todo.txt
@@ -1,15 +1,0 @@
-TODO LIST:
-
-- [x] make the drawining stay on the screen
-- [x] change players
-- [x] draw the players
-- [] melhorar a bola
-- [] procurar como fazer pro jogador nao sobreescrever o outro
-- [] fazer as condições de vitoria PARA CADA JOGADOR
-        - [x] diagonal
-        - [x] horizontal
-        - [x] vertical
-    - [x] determinar quem ganhou
-
-- [x] comment the code (comment a lot)
-- [] desenhar a linha caso alguem vença
