--- teste.py
@@ -1,1 +1,0 @@
-array = ['', '', '']
